refactor: Changed folder name from Web to Controller
